Lot's of scripts for data analysis

+++ launch_njst.sh
@@ -7,6 +7,7 @@
 #$ -v PATH
 
 BIN=/home/dmallo/njstM/scripts/NJstM
+nDigits=5
 
 if [ $# -lt 1 ]
 then
@@ -26,7 +27,7 @@ fi
 dir=$1
 
 module load R/3.2.2_1
-id=$(printf "%05d" $id)
+id=$(printf "%0${nDigits}d" $id)
 
 if [[ ! -d $dir/$id ]]
 then
@@ -38,22 +39,29 @@ else
 fi
 
 cd $dir/$id
+
 for i in $(find . -regex ".*\/.*\.tree[0-9]*")
+#for i in $(find . -regex ".*\/o.*\.tree")
 do
 	file=$(echo $i | sed -e "s/\.\/\(.*\)\.tree\(.*\)/\1.\2/g" -e "s/\.$//")
-	cat $i |sed -e "s/:[^),]*//g" -e "s/)[0-9.]*//g" -e "s/[(,);]/ /g" -e 's/ /\'$'\n''/g' |sort|uniq|tail -n+2|sed "s/\(.*\)\_.*\_.*$/& \1/" > ${file}.mapping
+	if [[ ! -f ${file}.mapping ]]
+	then
+		cat $i |sed -e "s/:[^),]*//g" -e "s/)[0-9.]*//g" -e "s/[(,);]/ /g" -e 's/ /\'$'\n''/g' |sort|uniq|tail -n+2|sed "s/\(.*\)\_.*\_.*$/& \1/" > ${file}.mapping
+	fi
+
 	if [[ ! -f ${file}.lnjst ]]
 	then
-
-		Rscript $BIN/njstm.r $i ${file}.mapping liu ${file}.lnjst
+		rm -f ${file}.lnjst.time
+		/usr/bin/time -a -o ${file}.lnjst.time Rscript $BIN/njstm.r $i ${file}.mapping liu ${file}.lnjst
 	fi
 	if [[ ! -f ${file}.onjst ]]
 	then
-
-		Rscript $BIN/njstm.r $i ${file}.mapping original ${file}.onjst
+		rm -f ${file}.onjst.time
+		/usr/bin/time -a -o ${file}.onjst.time Rscript $BIN/njstm.r $i ${file}.mapping original ${file}.onjst
 	fi
 	if [[ ! -f ${file}.unjst ]]
 	then 
-		Rscript $BIN/njstm.r $i ${file}.mapping unweighted ${file}.unjst
+		rm -f ${file}.onjst.time
+		/usr/bin/time -a -o ${file}.unjst.time Rscript $BIN/njstm.r $i ${file}.mapping unweighted ${file}.unjst
 	fi
 done
